requirements.txt
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.10.1
django-admin-honeypot==1.0.0
django-allauth==0.27.0
django-material==0.8.0
djangorestframework==3.4.7
gunicorn==19.6.0
oauthlib==2.0.0
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
wheel==0.24.0
whitenoise==3.2.1
psycopg2==2.4.5
